fix: fixing openApi and idtestata vals

+++ gdporch/src/main/java/it/csipiemonte/gdp/gdporch/model/service/GdpSospensioneService.java
@@ -5,6 +5,6 @@ import it.csipiemonte.gdp.gdporch.dto.SospensioneResponse;
 
 public interface GdpSospensioneService {
 
-    SospensioneResponse sospendi(Long idTestata, DateRangeRequest request);
+    SospensioneResponse sospendi(Integer idTestata, DateRangeRequest request);
 
 }
